requirements.txt
Sphinx==1.4.9
celery[redis]==3.1.25
docutils==0.12
futures==3.0.4
hiredis==0.2.0
jinja2
netaddr==0.7.18
pycares==1.0.0
pymongo==2.8.1
python-jenkins==0.4.11
python-dateutil
pytz
pyyaml==3.12
redis==2.10.6
requests
scandir
simplejson
sphinx-bootstrap-theme
sphinxcontrib-httpdomain==1.3.0
tornado==3.2.2
